Fix: sysctl parameters variable size u8 -> uint

+++ kernel/sched/fair.c
@@ -138,7 +138,7 @@ u8   __read_mostly sched_burst_fork_atavistic   = 2;
 u8   __read_mostly sched_burst_penalty_offset   = 22;
 uint __read_mostly sched_burst_penalty_scale    = 1280;
 uint __read_mostly sched_burst_cache_lifetime   = 60000000;
-static u8   sixty_four     = 64;
+static uint sixty_four     = 64;
 static uint maxval_12_bits = 4095;
 
 #define MAX_BURST_PENALTY (39U <<2)
